+++ kinode/src/kernel/mod.rs
@@ -13,14 +13,10 @@ use wasmtime::{Config, Engine, WasmBacktraceDetails};
 
 /// Manipulate a single process.
 pub mod process;
-/// Implement the functions served to processes by `wit-v0.7.0/kinode.wit`.
-mod standard_host;
-/// Implement the functions served to processes by `wit-v0.8.0/kinode.wit`.
-mod standard_host_v0;
 /// Implement the functions served to processes by `wit-v1.0.0/kinode.wit`.
 mod standard_host_v1;
 
-pub const LATEST_WIT_VERSION: u32 = 0;
+pub const LATEST_WIT_VERSION: u32 = 1;
 const PROCESS_CHANNEL_CAPACITY: usize = 100;
 
 #[derive(Serialize, Deserialize)]
