+++ kinode/src/kernel/process.rs
@@ -1,5 +1,5 @@
 use crate::KERNEL_PROCESS_ID;
-use lib::{types::core as t, v0::ProcessV0, v1::ProcessV1, Process};
+use lib::{types::core as t, v1::ProcessV1};
 use std::{
     collections::{HashMap, VecDeque},
     path::PathBuf,
@@ -54,37 +54,6 @@ pub struct ProcessState {
     pub message_queue: VecDeque<Result<t::KernelMessage, t::WrappedSendError>>,
     /// pipe for getting info about capabilities
     pub caps_oracle: t::CapMessageSender,
-}
-
-pub struct ProcessWasi {
-    pub process: ProcessState,
-    table: Table,
-    wasi: WasiCtx,
-}
-
-impl WasiView for ProcessWasi {
-    fn table(&mut self) -> &mut Table {
-        &mut self.table
-    }
-    fn ctx(&mut self) -> &mut WasiCtx {
-        &mut self.wasi
-    }
-}
-
-/// **can be removed in 1.0.0**
-pub struct ProcessWasiV0 {
-    pub process: ProcessState,
-    table: Table,
-    wasi: WasiCtx,
-}
-
-impl WasiView for ProcessWasiV0 {
-    fn table(&mut self) -> &mut Table {
-        &mut self.table
-    }
-    fn ctx(&mut self) -> &mut WasiCtx {
-        &mut self.wasi
-    }
 }
 
 pub struct ProcessWasiV1 {
@@ -146,94 +115,6 @@ async fn make_table_and_wasi(
     }
 
     (table, wasi.stderr(wasi_stderr.clone()).build(), wasi_stderr)
-}
-
-/// **can be removed in 1.0.0**
-async fn make_component(
-    engine: Engine,
-    wasm_bytes: &[u8],
-    home_directory_path: PathBuf,
-    process_state: ProcessState,
-) -> anyhow::Result<(Process, Store<ProcessWasi>, MemoryOutputPipe)> {
-    let component =
-        Component::new(&engine, wasm_bytes.to_vec()).expect("make_component: couldn't read file");
-
-    let mut linker = Linker::new(&engine);
-    Process::add_to_linker(&mut linker, |state: &mut ProcessWasi| state).unwrap();
-    let (table, wasi, wasi_stderr) = make_table_and_wasi(home_directory_path, &process_state).await;
-    wasmtime_wasi::add_to_linker_async(&mut linker).unwrap();
-
-    let our_process_id = process_state.metadata.our.process.clone();
-    let send_to_terminal = process_state.send_to_terminal.clone();
-
-    let mut store = Store::new(
-        &engine,
-        ProcessWasi {
-            process: process_state,
-            table,
-            wasi,
-        },
-    );
-
-    let bindings = match Process::instantiate_async(&mut store, &component, &linker).await {
-        Ok(b) => b,
-        Err(e) => {
-            t::Printout::new(
-                0,
-                t::KERNEL_PROCESS_ID.clone(),
-                format!("kernel: process {our_process_id} failed to instantiate: {e:?}"),
-            )
-            .send(&send_to_terminal)
-            .await;
-            return Err(e);
-        }
-    };
-
-    Ok((bindings, store, wasi_stderr))
-}
-
-/// **can be removed in 1.0.0**
-async fn make_component_v0(
-    engine: Engine,
-    wasm_bytes: &[u8],
-    home_directory_path: PathBuf,
-    process_state: ProcessState,
-) -> anyhow::Result<(ProcessV0, Store<ProcessWasiV0>, MemoryOutputPipe)> {
-    let component =
-        Component::new(&engine, wasm_bytes.to_vec()).expect("make_component: couldn't read file");
-
-    let mut linker = Linker::new(&engine);
-    ProcessV0::add_to_linker(&mut linker, |state: &mut ProcessWasiV0| state).unwrap();
-    let (table, wasi, wasi_stderr) = make_table_and_wasi(home_directory_path, &process_state).await;
-    wasmtime_wasi::add_to_linker_async(&mut linker).unwrap();
-
-    let our_process_id = process_state.metadata.our.process.clone();
-    let send_to_terminal = process_state.send_to_terminal.clone();
-
-    let mut store = Store::new(
-        &engine,
-        ProcessWasiV0 {
-            process: process_state,
-            table,
-            wasi,
-        },
-    );
-
-    let bindings = match ProcessV0::instantiate_async(&mut store, &component, &linker).await {
-        Ok(b) => b,
-        Err(e) => {
-            t::Printout::new(
-                0,
-                t::KERNEL_PROCESS_ID.clone(),
-                format!("kernel: process {our_process_id} failed to instantiate: {e:?}"),
-            )
-            .send(&send_to_terminal)
-            .await;
-            return Err(e);
-        }
-    };
-
-    Ok((bindings, store, wasi_stderr))
 }
 
 async fn make_component_v1(
@@ -347,83 +228,7 @@ pub async fn make_process_loop(
 
     let metadata = match wit_version {
         // assume missing version is oldest wit version
-        // **can be removed in 1.0.0**
-        None => {
-            let (bindings, mut store, wasi_stderr) =
-                make_component(engine, &wasm_bytes, home_directory_path, process_state).await?;
-
-            // the process will run until it returns from init() or crashes
-            match bindings.call_init(&mut store, &our.to_string()).await {
-                Ok(()) => {
-                    t::Printout::new(
-                        1,
-                        t::KERNEL_PROCESS_ID.clone(),
-                        format!("process {our} returned without error"),
-                    )
-                    .send(&send_to_terminal)
-                    .await;
-                }
-                Err(e) => {
-                    let stderr = wasi_stderr.contents().into();
-                    let stderr = String::from_utf8(stderr)?;
-                    let output = if !stderr.is_empty() {
-                        stderr
-                    } else {
-                        format!("{}", e.root_cause())
-                    };
-                    let error_text = if output.is_empty() {
-                        format!("\x1b[38;5;196mprocess {our} ended with error\x1b[0m")
-                    } else {
-                        format!("\x1b[38;5;196mprocess {our} ended with error:\x1b[0m\n{output}")
-                    };
-                    t::Printout::new(0, t::KERNEL_PROCESS_ID.clone(), error_text)
-                        .send(&send_to_terminal)
-                        .await;
-                }
-            };
-
-            // update metadata to what was mutated by process in store
-            store.data().process.metadata.to_owned()
-        }
-        // match version numbers
-        // **can be removed in 1.0.0**
-        Some(0) => {
-            let (bindings, mut store, wasi_stderr) =
-                make_component_v0(engine, &wasm_bytes, home_directory_path, process_state).await?;
-
-            // the process will run until it returns from init() or crashes
-            match bindings.call_init(&mut store, &our.to_string()).await {
-                Ok(()) => {
-                    t::Printout::new(
-                        1,
-                        t::KERNEL_PROCESS_ID.clone(),
-                        format!("process {our} returned without error"),
-                    )
-                    .send(&send_to_terminal)
-                    .await;
-                }
-                Err(e) => {
-                    let stderr = wasi_stderr.contents().into();
-                    let stderr = String::from_utf8(stderr)?;
-                    let output = if stderr != String::new() {
-                        stderr
-                    } else {
-                        format!("{}", e.root_cause())
-                    };
-                    t::Printout::new(
-                        0,
-                        t::KERNEL_PROCESS_ID.clone(),
-                        format!("\x1b[38;5;196mprocess {our} ended with error:\x1b[0m\n{output}"),
-                    )
-                    .send(&send_to_terminal)
-                    .await;
-                }
-            };
-
-            // update metadata to what was mutated by process in store
-            store.data().process.metadata.to_owned()
-        }
-        Some(1) | _ => {
+        None | Some(1) | _ => {
             let (bindings, mut store, wasi_stderr) =
                 make_component_v1(engine, &wasm_bytes, home_directory_path, process_state).await?;
