+++ kinode/src/kernel/standard_host_v1.rs
@@ -23,18 +23,18 @@ async fn print_debug(proc: &process::ProcessState, content: &str) {
 impl process::ProcessState {
     /// Ingest latest message directed to this process, and save it as the current message.
     /// If there is no message in the queue, wait async until one is received.
-    async fn get_next_message_for_process_v1(
+    async fn get_next_message_for_process(
         &mut self,
     ) -> Result<(wit::Address, wit::Message), (wit::SendError, Option<wit::Context>)> {
         let res = match self.message_queue.pop_front() {
             Some(message_from_queue) => message_from_queue,
-            None => self.ingest_message_v1().await,
+            None => self.ingest_message().await,
         };
-        self.kernel_message_to_process_receive_v1(res)
+        self.kernel_message_to_process_receive(res)
     }
 
     /// instead of ingesting latest, wait for a specific ID and queue all others
-    async fn get_specific_message_for_process_v1(
+    async fn get_specific_message_for_process(
         &mut self,
         awaited_message_id: u64,
     ) -> Result<(wit::Address, wit::Message), (wit::SendError, Option<wit::Context>)> {
@@ -43,20 +43,20 @@ impl process::ProcessState {
             match message {
                 Ok(ref km) if km.id == awaited_message_id => {
                     let km = self.message_queue.remove(i).unwrap();
-                    return self.kernel_message_to_process_receive_v1(km);
+                    return self.kernel_message_to_process_receive(km);
                 }
                 _ => continue,
             }
         }
         // next, wait for the awaited message to arrive
         loop {
-            let res = self.ingest_message_v1().await;
+            let res = self.ingest_message().await;
             let id = match &res {
                 Ok(km) => km.id,
                 Err(e) => e.id,
             };
             if id == awaited_message_id {
-                return self.kernel_message_to_process_receive_v1(res);
+                return self.kernel_message_to_process_receive(res);
             } else {
                 self.message_queue.push_back(res);
             }
@@ -66,7 +66,7 @@ impl process::ProcessState {
     /// ingest next valid message from kernel.
     /// cancel any timeout task associated with this message.
     /// if the message is a response, only enqueue if we have an outstanding request for it.
-    async fn ingest_message_v1(&mut self) -> Result<t::KernelMessage, t::WrappedSendError> {
+    async fn ingest_message(&mut self) -> Result<t::KernelMessage, t::WrappedSendError> {
         loop {
             let message = self
                 .recv_in_process
@@ -98,7 +98,7 @@ impl process::ProcessState {
 
     /// Convert a message from the main event loop into a result for the process to receive.
     /// If the message is a response or error, get context if we have one.
-    fn kernel_message_to_process_receive_v1(
+    fn kernel_message_to_process_receive(
         &mut self,
         incoming: Result<t::KernelMessage, t::WrappedSendError>,
     ) -> Result<(wit::Address, wit::Message), (wit::SendError, Option<wit::Context>)> {
@@ -149,10 +149,10 @@ impl process::ProcessState {
             Err(e) => {
                 self.last_message_blobbed = false;
                 match self.contexts.remove(&e.id) {
-                    None => return Err((t::en_wit_send_error_v1(e.error), None)),
+                    None => return Err((t::en_wit_send_error(e.error), None)),
                     Some((context, _timeout_handle)) => {
                         self.prompting_message = context.prompting_message;
-                        return Err((t::en_wit_send_error_v1(e.error), context.context));
+                        return Err((t::en_wit_send_error(e.error), context.context));
                     }
                 }
             }
@@ -201,14 +201,12 @@ impl process::ProcessState {
         };
 
         Ok((
-            km.source.en_wit_v1(),
+            km.source.en_wit(),
             match km.message {
-                t::Message::Request(request) => {
-                    wit::Message::Request(t::en_wit_request_v1(request))
-                }
+                t::Message::Request(request) => wit::Message::Request(t::en_wit_request(request)),
                 // NOTE: we throw away whatever context came from the sender, that's not ours
                 t::Message::Response((response, _sent_context)) => {
-                    wit::Message::Response((t::en_wit_response_v1(response), context))
+                    wit::Message::Response((t::en_wit_response(response), context))
                 }
             },
         ))
@@ -218,7 +216,7 @@ impl process::ProcessState {
     /// will only fail if process does not have capability to send to target.
     /// if the request has a timeout (expects response), start a task to track
     /// that timeout and return timeout error if it expires.
-    async fn send_request_v1(
+    async fn send_request(
         &mut self,
         // only used when kernel steps in to get/set state
         fake_source: Option<t::Address>,
@@ -228,7 +226,7 @@ impl process::ProcessState {
         blob: Option<wit::LazyLoadBlob>,
     ) -> Result<u64> {
         let source = fake_source.unwrap_or(self.metadata.our.clone());
-        let mut request = t::de_wit_request_v1(request);
+        let mut request = t::de_wit_request(request);
 
         // if request chooses to inherit, it means to take the ID and lazy_load_blob,
         // if any, from the last message it ingested
@@ -287,7 +285,7 @@ impl process::ProcessState {
             let this_request = request.clone();
             let this_blob = blob.clone();
             let self_sender = self.self_sender.clone();
-            let original_target = t::Address::de_wit_v1(target.clone());
+            let original_target = t::Address::de_wit(target.clone());
             let timeout_handle = tokio::spawn(async move {
                 tokio::time::sleep(std::time::Duration::from_secs(timeout_secs)).await;
                 let _ = self_sender
@@ -322,7 +320,7 @@ impl process::ProcessState {
         let kernel_message = t::KernelMessage {
             id: request_id,
             source,
-            target: t::Address::de_wit_v1(target),
+            target: t::Address::de_wit(target),
             rsvp: match (
                 request.expects_response,
                 request.inherit,
@@ -352,8 +350,8 @@ impl process::ProcessState {
     }
 
     /// takes Response generated by a process and sends it to the main event loop.
-    async fn send_response_v1(&mut self, response: wit::Response, blob: Option<wit::LazyLoadBlob>) {
-        let mut response = t::de_wit_response_v1(response);
+    async fn send_response(&mut self, response: wit::Response, blob: Option<wit::LazyLoadBlob>) {
+        let mut response = t::de_wit_response(response);
 
         // the process requires a prompting_message in order to issue a response
         let Some(ref prompting_message) = self.prompting_message else {
@@ -379,7 +377,7 @@ impl process::ProcessState {
 
         let blob = match response.inherit {
             true => self.last_blob.clone(),
-            false => t::de_wit_blob_v1(blob),
+            false => t::de_wit_blob(blob),
         };
 
         if !response.capabilities.is_empty() {
@@ -433,7 +431,7 @@ async fn send_and_await_response(
             process.process.metadata.our.process
         ));
     }
-    if t::Address::de_wit_v1(target.clone()) == process.process.metadata.our {
+    if t::Address::de_wit(target.clone()) == process.process.metadata.our {
         return Err(anyhow::anyhow!(
             "kernel: got invalid send_and_await_response() Request from and to {}: cannot await a Request to `our`: will deadlock",
             process.process.metadata.our,
@@ -441,14 +439,10 @@ async fn send_and_await_response(
     }
     let id = process
         .process
-        .send_request_v1(source, target, request, None, blob)
+        .send_request(source, target, request, None, blob)
         .await;
     match id {
-        Ok(id) => match process
-            .process
-            .get_specific_message_for_process_v1(id)
-            .await
-        {
+        Ok(id) => match process.process.get_specific_message_for_process(id).await {
             Ok((address, wit::Message::Response(response))) => {
                 Ok(Ok((address, wit::Message::Response(response))))
             }
@@ -490,7 +484,7 @@ impl StandardHost for process::ProcessWasiV1 {
     }
 
     async fn our(&mut self) -> Result<wit::Address> {
-        Ok(self.process.metadata.our.en_wit_v1())
+        Ok(self.process.metadata.our.en_wit())
     }
 
     //
@@ -498,18 +492,18 @@ impl StandardHost for process::ProcessWasiV1 {
     //
 
     async fn set_on_exit(&mut self, on_exit: wit::OnExit) -> Result<()> {
-        let on_exit = t::OnExit::de_wit_v1(on_exit);
+        let on_exit = t::OnExit::de_wit(on_exit);
         self.process.metadata.on_exit = on_exit.clone();
         match self
             .process
-            .send_request_v1(
+            .send_request(
                 Some(t::Address {
                     node: self.process.metadata.our.node.clone(),
                     process: KERNEL_PROCESS_ID.clone(),
                 }),
                 wit::Address {
                     node: self.process.metadata.our.node.clone(),
-                    process: KERNEL_PROCESS_ID.en_wit_v1(),
+                    process: KERNEL_PROCESS_ID.en_wit(),
                 },
                 wit::Request {
                     inherit: false,
@@ -536,7 +530,7 @@ impl StandardHost for process::ProcessWasiV1 {
     }
 
     async fn get_on_exit(&mut self) -> Result<wit::OnExit> {
-        Ok(self.process.metadata.on_exit.en_wit_v1())
+        Ok(self.process.metadata.on_exit.en_wit())
     }
 
     /// create a message from the *kernel* to the filesystem,
@@ -551,7 +545,7 @@ impl StandardHost for process::ProcessWasiV1 {
             }),
             wit::Address {
                 node: self.process.metadata.our.node.clone(),
-                process: STATE_PROCESS_ID.en_wit_v1(),
+                process: STATE_PROCESS_ID.en_wit(),
             },
             wit::Request {
                 inherit: false,
@@ -597,7 +591,7 @@ impl StandardHost for process::ProcessWasiV1 {
             }),
             wit::Address {
                 node: self.process.metadata.our.node.clone(),
-                process: STATE_PROCESS_ID.en_wit_v1(),
+                process: STATE_PROCESS_ID.en_wit(),
             },
             wit::Request {
                 inherit: false,
@@ -638,7 +632,7 @@ impl StandardHost for process::ProcessWasiV1 {
             }),
             wit::Address {
                 node: self.process.metadata.our.node.clone(),
-                process: STATE_PROCESS_ID.en_wit_v1(),
+                process: STATE_PROCESS_ID.en_wit(),
             },
             wit::Request {
                 inherit: false,
@@ -683,7 +677,7 @@ impl StandardHost for process::ProcessWasiV1 {
         let old_last_blob = self.process.last_blob.clone();
         let vfs_address = wit::Address {
             node: self.process.metadata.our.node.clone(),
-            process: VFS_PROCESS_ID.en_wit_v1(),
+            process: VFS_PROCESS_ID.en_wit(),
         };
         let Ok(Ok((_, hash_response))) = send_and_await_response(
             self,
@@ -744,7 +738,7 @@ impl StandardHost for process::ProcessWasiV1 {
                     on: self.process.metadata.our.process.clone(),
                     caps: request_capabilities
                         .into_iter()
-                        .map(|cap| t::de_wit_capability_v1(cap).0)
+                        .map(|cap| t::de_wit_capability(cap).0)
                         .collect(),
                     responder: tx,
                 })
@@ -762,7 +756,7 @@ impl StandardHost for process::ProcessWasiV1 {
             }),
             wit::Address {
                 node: self.process.metadata.our.node.clone(),
-                process: KERNEL_PROCESS_ID.en_wit_v1(),
+                process: KERNEL_PROCESS_ID.en_wit(),
             },
             wit::Request {
                 inherit: false,
@@ -771,7 +765,7 @@ impl StandardHost for process::ProcessWasiV1 {
                     id: new_process_id.clone(),
                     wasm_bytes_handle: wasm_path,
                     wit_version: self.process.metadata.wit_version,
-                    on_exit: t::OnExit::de_wit_v1(on_exit),
+                    on_exit: t::OnExit::de_wit(on_exit),
                     initial_capabilities: request_capabilities_filtered
                         .into_iter()
                         .map(|(cap, _sig)| cap)
@@ -799,7 +793,7 @@ impl StandardHost for process::ProcessWasiV1 {
             self.process
                 .caps_oracle
                 .send(t::CapMessage::Add {
-                    on: t::ProcessId::de_wit_v1(process_id),
+                    on: t::ProcessId::de_wit(process_id),
                     caps: vec![t::Capability::new(
                         (self.process.metadata.our.node.clone(), &new_process_id),
                         params,
@@ -819,7 +813,7 @@ impl StandardHost for process::ProcessWasiV1 {
             }),
             wit::Address {
                 node: self.process.metadata.our.node.clone(),
-                process: KERNEL_PROCESS_ID.en_wit_v1(),
+                process: KERNEL_PROCESS_ID.en_wit(),
             },
             wit::Request {
                 inherit: false,
@@ -874,7 +868,7 @@ impl StandardHost for process::ProcessWasiV1 {
             .unwrap();
         rx.await.unwrap();
         print_debug(&self.process, "spawned a new process").await;
-        Ok(Ok(new_process_id.en_wit_v1().to_owned()))
+        Ok(Ok(new_process_id.en_wit().to_owned()))
     }
 
     //
@@ -890,7 +884,7 @@ impl StandardHost for process::ProcessWasiV1 {
                 on: self.process.metadata.our.process.clone(),
                 caps: caps
                     .iter()
-                    .map(|cap| t::de_wit_capability_v1(cap.clone()).0)
+                    .map(|cap| t::de_wit_capability(cap.clone()).0)
                     .collect(),
                 responder: Some(tx),
             })
@@ -908,7 +902,7 @@ impl StandardHost for process::ProcessWasiV1 {
                 on: self.process.metadata.our.process.clone(),
                 caps: caps
                     .iter()
-                    .map(|cap| t::de_wit_capability_v1(cap.clone()).0)
+                    .map(|cap| t::de_wit_capability(cap.clone()).0)
                     .collect(),
                 responder: Some(tx),
             })
@@ -930,7 +924,7 @@ impl StandardHost for process::ProcessWasiV1 {
         let caps = rx.await?;
         Ok(caps
             .into_iter()
-            .map(|cap| t::en_wit_capability_v1(cap))
+            .map(|cap| t::en_wit_capability(cap))
             .collect())
     }
 
@@ -943,7 +937,7 @@ impl StandardHost for process::ProcessWasiV1 {
     async fn receive(
         &mut self,
     ) -> Result<Result<(wit::Address, wit::Message), (wit::SendError, Option<wit::Context>)>> {
-        Ok(self.process.get_next_message_for_process_v1().await)
+        Ok(self.process.get_next_message_for_process().await)
     }
 
     /// from a process: check if the last message received had a blob.
@@ -955,7 +949,7 @@ impl StandardHost for process::ProcessWasiV1 {
     /// if the last message did not have a blob, will return None.
     async fn get_blob(&mut self) -> Result<Option<wit::LazyLoadBlob>> {
         Ok(if self.process.last_message_blobbed {
-            t::en_wit_blob_v1(self.process.last_blob.clone())
+            t::en_wit_blob(self.process.last_blob.clone())
         } else {
             None
         })
@@ -964,7 +958,7 @@ impl StandardHost for process::ProcessWasiV1 {
     /// from a process: grab the **most recent** blob that has ever been received.
     /// if no blobs have ever been received, will return None.
     async fn last_blob(&mut self) -> Result<Option<wit::LazyLoadBlob>> {
-        Ok(t::en_wit_blob_v1(self.process.last_blob.clone()))
+        Ok(t::en_wit_blob(self.process.last_blob.clone()))
     }
 
     async fn send_request(
@@ -976,7 +970,7 @@ impl StandardHost for process::ProcessWasiV1 {
     ) -> Result<()> {
         let id = self
             .process
-            .send_request_v1(None, target, request, context, blob)
+            .send_request(None, target, request, context, blob)
             .await;
         match id {
             Ok(_id) => Ok(()),
@@ -996,7 +990,7 @@ impl StandardHost for process::ProcessWasiV1 {
         for request in requests {
             let id = self
                 .process
-                .send_request_v1(None, request.0, request.1, request.2, request.3)
+                .send_request(None, request.0, request.1, request.2, request.3)
                 .await;
             match id {
                 Ok(_id) => continue,
@@ -1011,7 +1005,7 @@ impl StandardHost for process::ProcessWasiV1 {
         response: wit::Response,
         blob: Option<wit::LazyLoadBlob>,
     ) -> Result<()> {
-        self.process.send_response_v1(response, blob).await;
+        self.process.send_response(response, blob).await;
         Ok(())
     }
