+++ kinode/src/terminal/utils.rs
@@ -61,14 +61,14 @@ pub fn splash(
                >^#########@      888   Y88b  888 888  888 Y88..88P Y88b 888 Y8b.
                  `>#######`      888    Y88b 888 888  888  "Y88P"   "Y88888  "Y8888
                 .>######%
-               /###%^#%          {} ({})
-             /##%@#  `           runtime version {}
-          ./######`              a general purpose sovereign cloud computer
+               /###%^#%       {} ({})
+             /##%@#  `        runtime version {}
+          ./######`           a general purpose sovereign cloud computer
         /.^`.#^#^`
-       `   ,#`#`#,               public IP {}
-          ,/ /` `                home directory at {}
+       `   ,#`#`#,          public IP {}
+          ,/ /` `           home directory at {}
         .*`
- networking public key: {}{}
+ networking pubkey: {}{}
 "#,
                 our.name,
                 if our.is_direct() {
