+++ lib/build.rs
@@ -4,12 +4,6 @@ use std::path::Path;
 use sha2::{Digest, Sha256};
 
 const KIT_CACHE: &str = "/tmp/kinode-kit-cache";
-// TODO remove for 1.0
-const KINODE_WIT_0_7_0_URL: &str =
-    "https://raw.githubusercontent.com/kinode-dao/kinode-wit/aa2c8b11c9171b949d1991c32f58591c0e881f85/kinode.wit";
-// TODO remove for 1.0
-const KINODE_WIT_0_8_0_URL: &str =
-    "https://raw.githubusercontent.com/kinode-dao/kinode-wit/v0.8/kinode.wit";
 const KINODE_WIT_1_0_0_URL: &str =
     "https://raw.githubusercontent.com/kinode-dao/kinode-wit/v1.0.0/kinode.wit";
 
@@ -64,24 +58,6 @@ fn main() {
     }
 
     let pwd = std::env::current_dir().expect("Failed to get current directory");
-
-    let wit_file = pwd.join("wit-v0.7.0").join("kinode.wit");
-
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    rt.block_on(async {
-        download_file(KINODE_WIT_0_7_0_URL, &wit_file)
-            .await
-            .expect("Failed to download WIT 0.7");
-    });
-
-    let wit_file = pwd.join("wit-v0.8.0").join("kinode.wit");
-
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    rt.block_on(async {
-        download_file(KINODE_WIT_0_8_0_URL, &wit_file)
-            .await
-            .expect("Failed to download WIT 0.8");
-    });
 
     let wit_file = pwd.join("wit-v1.0.0").join("kinode.wit");
