+++ lib/src/lib.rs
@@ -17,9 +17,6 @@ pub mod types {
     pub use crate::http::server_types as http_server;
 }
 
-pub use kinode::process;
-pub use kinode::process::standard as wit;
-
 // `trappable_imports: true` keeps behavior the same as pre-240410
 //  where imports are wrapped with an `anyhow::Result`.
 //  This allows errors that occur in imports to be handled naturally,
@@ -28,28 +25,6 @@ pub use kinode::process::standard as wit;
 //
 //  source:
 //  https://github.com/bytecodealliance/wasmtime/commit/1cf0060bbc17aaf35b81b989c6394df254bb4f2e
-
-// can remove in 1.0!
-
-wasmtime::component::bindgen!({
-    path: "wit-v0.7.0",
-    world: "process",
-    async: true,
-    trappable_imports: true,
-});
-
-// can remove in 1.0!
-
-pub mod v0 {
-    pub use kinode::process;
-    pub use kinode::process::standard as wit;
-    wasmtime::component::bindgen!({
-        path: "wit-v0.8.0",
-        world: "process-v0",
-        async: true,
-        trappable_imports: true,
-    });
-}
 
 pub mod v1 {
     pub use kinode::process;
